Adjust motor currents
From Klipper Discord:

Coffee[CoffeeMaker,Chonxy]03/01/2020
Quick explanation of current setting on tmc drivers. the value you put in shuld be the RMS value. not peak value that is written on the stepper.
The calculation of rms from peak current is this:

Motor_Peak_Current x 0.707

And since we dont wanna push the motors with 100% we do:

Motor_Peak_Current x 0.707 x 0.80

that will make a 20% safety margin.

so a stock ender 3 motor that is 0.84A peak

0.84x0.707x0.8 = 0.47A

so you put 0.47 in run_current of your tmc stepper.

A normal 1.4A stepper is:

1.4x0.707x0.8 = 0.79A

so you put 0.79 in run_current of your tmc stepper.

the Hold current there is no "rule" about. i like to do Run_Current - 0.2
but that will depend on the use case.